fix casting (#11)

+++ liblangutil/ErrorReporter.cpp
@@ -1,4 +1,4 @@
-/*
+/*error_reporter
 	This file is part of solidity.
 
 	solidity is free software: you can redistribute it and/or modify
@@ -162,8 +162,8 @@ void ErrorReporter::removeError(ErrorId _errorId)
     );
     if (it != errors.end())
     {
-        m_errorCount -= std::distance(it, errors.end());
-        errors.erase(it, errors.end());
+		m_errorCount -= static_cast<size_t>(std::distance(it, errors.end()));
+		errors.erase(it, errors.end());
     }
 }
